first commit of 2024

+++ controllers/cards.js
@@ -1,43 +1,46 @@
 const CardModel = require('../models/card');
+const BadRequestError = require('../utils/errors/bad-request');
+const ForbiddenError = require('../utils/errors/forbidden');
+const NotFoundError = require('../utils/errors/not-found');
+const { setStatusCreated } = require('../utils/statusSetter');
 
-const {
-  setStatusCreated,
-  setStatusNotFound,
-  setStatusBadRequest,
-  setStatusServerError,
-} = require('../utils/statusSetter');
-
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const cardData = req.body;
   cardData.owner = req.user._id;
   return CardModel.create(cardData)
     .then((data) => setStatusCreated(res, data))
     .catch((err) => {
-      if (err.name === 'ValidationError') return setStatusBadRequest(res);
-      return setStatusServerError(res);
-    });
+      if (err.name === 'ValidationError') throw new BadRequestError();
+      return next(err);
+    })
+    .catch(next);
 };
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   CardModel.find()
     .then((cards) => res.send(cards))
-    .catch(() => setStatusServerError(res));
+    .catch(next);
 };
 
-module.exports.deleteCardById = (req, res) => {
+module.exports.deleteCardById = (req, res, next) => {
   const { cardId } = req.params;
-  CardModel.findByIdAndRemove(cardId)
+  CardModel.findById(cardId)
     .then((card) => {
-      if (!card) return setStatusNotFound(res);
-      return res.send({ message: 'Card Deleted' });
+      if (!card) throw new NotFoundError('Card not found');
+      const cardOwnerId = card.owner.toString();
+      if (cardOwnerId !== req.user._id) throw new ForbiddenError();
+      return CardModel.findByIdAndRemove(cardId)
+        .then(() => res.send({ message: 'Card Deleted' }))
+        .catch(next);
     })
     .catch((err) => {
-      if (err.name === 'CastError') return setStatusBadRequest(res);
-      return setStatusServerError(res);
-    });
+      if (err.name === 'CastError') throw new BadRequestError();
+      return next(err);
+    })
+    .catch(next);
 };
 
-module.exports.setCardLike = (req, res) => {
+module.exports.setCardLike = (req, res, next) => {
   const { cardId } = req.params;
   return CardModel.findByIdAndUpdate(
     cardId,
@@ -45,16 +48,17 @@ module.exports.setCardLike = (req, res) => {
     { new: true },
   )
     .then((card) => {
-      if (!card) return setStatusNotFound(res);
+      if (!card) throw new NotFoundError('Card not found');
       return res.send(card.likes);
     })
     .catch((err) => {
-      if (err.name === 'CastError') return setStatusBadRequest(res);
-      return setStatusServerError(res);
-    });
+      if (err.name === 'CastError') throw new BadRequestError();
+      return next(err);
+    })
+    .catch(next);
 };
 
-module.exports.removeCardLike = (req, res) => {
+module.exports.removeCardLike = (req, res, next) => {
   const { cardId } = req.params;
   return CardModel.findByIdAndUpdate(
     cardId,
@@ -62,11 +66,12 @@ module.exports.removeCardLike = (req, res) => {
     { new: true },
   )
     .then((card) => {
-      if (!card) return setStatusNotFound(res);
+      if (!card) throw new NotFoundError('Card not found');
       return res.send(card.likes);
     })
     .catch((err) => {
-      if (err.name === 'CastError') return setStatusBadRequest(res);
-      return setStatusServerError(res);
-    });
+      if (err.name === 'CastError') throw new BadRequestError();
+      return next(err);
+    })
+    .catch(next);
 };
